PLAT-194: add handlers for all methods
